create pvlib\tests\spectrum\conftest.py
